+++ sat/cli/bootsys/parser.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2020-2022 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2020-2023 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -137,6 +137,18 @@ def _add_bos_template_options(subparser, action):
              'will be expanded recursively into their constituent node xnames. For '
              'instance, if a slot xname is given as part of --bos-limit, then that '
              'xname will be expanded into the node xnames for all nodes in that slot.'
+    )
+
+    subparser.add_argument(
+        '--staged-session', action='store_true',
+        help=f'If specified, then create a "staged" BOS session. A staged session differs '
+             f'from a normal BOS session in that a staged session does not automatically '
+             f'change the state of any components targeted by the session templates. '
+             f'Instead, a staged session will update components\' "staged_state", which '
+             f'can later be applied through the command \'cray bos v2 applystaged create\', '
+             f'or through an API call to BOS. When using this option, the command will '
+             f'not wait for completion of BOS sessions. This option cannot be used with '
+             f'BOS v1.'
     )